requirements.txt
requests
rapidfuzz
python-dotenv
ollama